Major refactor with an eye on v1.0.
The goal of the refactor is to make namespaces and actions more intuitive to
use. Once declared, namespaces can be accessed directly as properties of
virgilio. Actions can be invoked like functions.

A second goal was to reduce to complexity of Virgilio (even further). The
mediator was considered superfluous and removed, taking with it timeouts, pubsub
and, for now, forking. `execute()` is gone too, as there is no use for it
anymore.

Before joining these changes back to master, some examples/tests might need to
be rewritten to better display the new way of working (though the current
examples/tests work fine).

Also, a backwards-compatibility module should be written to add the features
that dissapeared back in for projects that depend on them.